PremiereEcoute.Sessions.Scores.Poll (Premiere Ecoute v0.1.0)

Poll schema and operations

Manages voting polls associated with listening sessions and tracks. This module provides data persistence for poll results, vote tallying, and validation to ensure vote counts are consistent with totals.

Types

entity(type)

@type entity(type) :: type | nil | Ecto.Association.NotLoaded.t()

nullable(type)

@type nullable(type) :: type | nil

t()

@type t() :: %PremiereEcoute.Sessions.Scores.Poll{
  __meta__: term(),
  ended_at: NaiveDateTime.t() | nil,
  id: integer() | nil,
  inserted_at: NaiveDateTime.t() | nil,
  poll_id: String.t() | nil,
  session: entity(PremiereEcoute.Sessions.ListeningSession.t()),
  session_id: term(),
  title: String.t() | nil,
  total_votes: integer() | nil,
  track: entity(PremiereEcoute.Sessions.Discography.Album.Track.t()),
  track_id: term(),
  updated_at: NaiveDateTime.t() | nil,
  votes: map() | nil
}

Functions

all(clauses \\ [])

all_by(query \\ __MODULE__, clauses)

average(query \\ __MODULE__, field)

changeset(poll, attrs)

count(query \\ __MODULE__, field)

create(entity)

create_if_not_exists(entity)

delete(entity)

delete_all(query \\ __MODULE__)

get(id)

get_by(query \\ __MODULE__, clauses)

max(query \\ __MODULE__, field)

min(query \\ __MODULE__, field)

page(clauses \\ [], page, page_size \\ 1)

preload(entity)

sum(query \\ __MODULE__, field)

update(entity, attrs)

upsert(poll)

@spec upsert(t()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

upsert(entity, attrs)