PremiereEcouteWeb.Layouts (Premiere Ecoute v0.1.0)

This module holds different layouts used by your application.

See the layouts directory for all templates available. The "root" layout is a skeleton rendered as part of the application router. The "app" layout is rendered as component in regular views and live views.

Functions

app(assigns)

Renders the app layout for our streaming dashboard

Examples

<Layouts.app flash={@flash}>
  <h1>Content</h1>
</Layouts.app>

Attributes

flash (:map) (required) - the map of flash messages.

current_scope (:map) - the current scope. Defaults to nil.

Slots

inner_block (required)

flash_group(assigns)

Shows the flash group with standard titles and content.

Examples

<.flash_group flash={@flash} />

Attributes

flash (:map) (required) - the map of flash messages.

id (:string) - the optional id of flash container. Defaults to "flash-group".

root(assigns)